Add new connection parameter for HANA Cloud
With this new parameter you can override the automatic detection of
whether you are connecting to a HANA Cloud or On-prem host. This
addresses the issue that in some cases the 'build_branch' value that is
returned by the HANA is empty and thus the On-prem SQL is used against
a HANA Cloud.